backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
nselib
yfinance
redis
gotrue
supabase
python-dotenv
pydantic-settings
# Explicitly allowing modern versions
httpx>=0.27.0
# pyotp - Removed as it was for Angel One
websocket-client
logzero
isodate
pandas_market_calendars
beautifulsoup4
lxml
requests
fake-useragent
